requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
APScheduler==3.10.4
jinja2==3.1.2
requests==2.31.0
pydantic==2.5.0